chore: Merge `InvitationController` into `InvitationService`
