+++ crates/rest-api/src/features/invitations/get_invitation.rs
@@ -6,7 +6,7 @@
 use std::ops::Deref as _;
 
 use rocket::{get, serde::json::Json};
-use service::invitations::{InvitationController, InvitationToken};
+use service::invitations::{InvitationService, InvitationToken};
 
 use crate::{
     error::{self, Error},
@@ -19,12 +19,12 @@ use super::{forms::InvitationTokenType, model::*};
 /// Get information about a workspace invitation.
 #[get("/v1/invitations/<invitation_id>")]
 pub async fn get_invitation_route<'r>(
-    invitation_controller: LazyGuard<InvitationController>,
+    invitation_service: LazyGuard<InvitationService>,
     invitation_id: i32,
 ) -> Result<Json<WorkspaceInvitation>, Error> {
-    let invitation_controller = invitation_controller.inner?;
+    let invitation_service = invitation_service.inner?;
 
-    let invitation = invitation_controller.get(&invitation_id).await?;
+    let invitation = invitation_service.get(&invitation_id).await?;
     let Some(invitation) = invitation else {
         return Err(Error::from(error::NotFound {
             reason: format!("No invitation with id '{invitation_id}'"),
@@ -38,16 +38,16 @@ pub async fn get_invitation_route<'r>(
 /// Get information about an invitation from an accept or reject token.
 #[get("/v1/invitation-tokens/<token>/details?<token_type>")]
 pub async fn get_invitation_token_details_route<'r>(
-    invitation_controller: LazyGuard<InvitationController>,
+    invitation_service: LazyGuard<InvitationService>,
     token: Uuid,
     token_type: InvitationTokenType,
 ) -> Result<Json<WorkspaceInvitationBasicDetails>, Error> {
-    let invitation_controller = invitation_controller.inner?;
+    let invitation_service = invitation_service.inner?;
     let token = InvitationToken::from(*token.deref());
 
     let invitation = match token_type {
-        InvitationTokenType::Accept => invitation_controller.get_by_accept_token(token).await,
-        InvitationTokenType::Reject => invitation_controller.get_by_reject_token(token).await,
+        InvitationTokenType::Accept => invitation_service.get_by_accept_token(token).await,
+        InvitationTokenType::Reject => invitation_service.get_by_reject_token(token).await,
     }?;
     let Some(invitation) = invitation else {
         return Err(error::Forbidden("No invitation found for provided token".to_string()).into());
