+++ crates/rest-api/src/features/invitations/guards/mod.rs
@@ -3,7 +3,4 @@
 // Copyright: 2024, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
-mod invitation_controller;
-mod unauthenticated_invitation_service;
-
-pub use unauthenticated_invitation_service::*;
+mod invitation_service;
