+++ crates/rest-api/src/features/members/guards/mod.rs
@@ -1,6 +1,7 @@
 // prose-pod-api
 //
-// Copyright: 2024, Rémi Bardon <remi@remibardon.name>
+// Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 mod member_service;
+mod unauthenticated_member_service;
