+++ crates/service/src/features/invitations/mod.rs
@@ -4,14 +4,12 @@
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 pub mod entities;
-pub mod invitation_controller;
 pub mod invitation_repository;
 pub mod invitation_service;
 pub(crate) mod migrations;
 pub mod models;
 
 pub use entities::*;
-pub use invitation_controller::*;
 pub use invitation_repository::*;
 pub use invitation_service::*;
 pub use models::*;
